# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlightMovement API Documentation",
        "description": "API documentation for FlightMovement",
        "contact": {
            "email": "dev@camranh.aero"
        },
        "version": "1.0.0"
    },
    "servers": [
        {

            "url": "http://localhost:8000/api",
            "description": "Rest API"
        }
    ],
    "paths": {
...
        "/airports": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get all active airports",
                "description": "Retrieve all active airports ordered by name",
                "operationId": "18e2a102579da344c6e74f5ed472011b",
                "responses": {
                    "200": {
                        "description": "Airports retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "AirportId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "Name": {
                                                        "type": "string",
                                                        "example": "Tan Son Nhat International Airport"
                                                    },
                                                    "IataCode": {
                                                        "type": "string",
                                                        "example": "SGN"
                                                    },
                                                    "IcaoCode": {
                                                        "type": "string",
                                                        "example": "VVTS"
                                                    },
                                                    "City": {
                                                        "type": "string",
                                                        "example": "Ho Chi Minh City"
                                                    },
                                                    "Country": {
                                                        "type": "string",
                                                        "example": "Vietnam"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Airports"
                ],
                "summary": "Create new airport",
                "description": "Create a new airport with the provided information",
                "operationId": "00dd59408628cb1464694d94c9d5cf8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "AirportId",
                                    "Name"
                                ],
                                "properties": {
                                    "AirportId": {
                                        "description": "Unique airport ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Name": {
                                        "description": "Airport name",
                                        "type": "string",
                                        "example": "Tan Son Nhat International Airport"
                                    },
                                    "IataCode": {
                                        "description": "IATA code",
                                        "type": "string",
                                        "example": "SGN"
                                    },
                                    "IcaoCode": {
                                        "description": "ICAO code",
                                        "type": "string",
                                        "example": "VVTS"
                                    },
                                    "City": {
                                        "description": "City name",
                                        "type": "string",
                                        "example": "Ho Chi Minh City"
                                    },
                                    "Country": {
                                        "description": "Country name",
                                        "type": "string",
                                        "example": "Vietnam"
                                    },
                                    "CustomsType": {
                                        "description": "Customs type",
                                        "type": "string",
                                        "example": "International"
                                    },
                                    "Status": {
                                        "description": "Status (0=inactive, 1=active)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Airport created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airport created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/airports/iata/{iataCode}": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get airport by IATA code",
                "description": "Retrieve airport information by IATA code",
                "operationId": "8ae170ef2485238a6350c33aea066db1",
                "parameters": [
                    {
                        "name": "iataCode",
                        "in": "path",
                        "description": "IATA code (e.g., SGN, HAN)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "SGN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Airport found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "AirportId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "Name": {
                                                    "type": "string",
                                                    "example": "Tan Son Nhat International Airport"
                                                },
                                                "IataCode": {
                                                    "type": "string",
                                                    "example": "SGN"
                                                },
                                                "IcaoCode": {
                                                    "type": "string",
                                                    "example": "VVTS"
                                                },
                                                "City": {
                                                    "type": "string",
                                                    "example": "Ho Chi Minh City"
                                                },
                                                "Country": {
                                                    "type": "string",
                                                    "example": "Vietnam"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Airport not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airport not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/airports/city/{city}": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get airports by city",
                "description": "Retrieve all airports in a specific city",
                "operationId": "a8e20f83b93ddf2bf18692a20ef94976",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "City name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Ho Chi Minh City"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Airports retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "AirportId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "Name": {
                                                        "type": "string",
                                                        "example": "Tan Son Nhat International Airport"
                                                    },
                                                    "IataCode": {
                                                        "type": "string",
                                                        "example": "SGN"
                                                    },
                                                    "IcaoCode": {
                                                        "type": "string",
                                                        "example": "VVTS"
                                                    },
                                                    "City": {
                                                        "type": "string",
                                                        "example": "Ho Chi Minh City"
                                                    },
                                                    "Country": {
                                                        "type": "string",
                                                        "example": "Vietnam"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}